Added request token test for when there is a message and amount is specifically 0
