+++ e2e/fixtures/page.mjs
@@ -1,3 +1,3 @@
 export const fixture = {
-  page: undefined,
+  page: null,
 }
